* Moved `chunkUpload` backend code to utils file
