+++ webpack.config.js
@@ -4,67 +4,11 @@ const webpack           = require('webpack')
 
 const packageInfo       = require('./package')
 
-const bodyParser        = require('webpack-body-parser')
+const chunkUpload      = require('./src/utils/chunkUpload')
 
 process.env.NODE_ENV = process.env.NODE_ENV || 'production'
 
 const isDev = process.env.NODE_ENV === 'development'
-
-const CHUNK_SIZE = 1048576
-const ChunkActiveUploads = {}
-
-const chunkUploadStart = (req, res) => {
-  const uuid = Math.floor((1 + Math.random()) * 0x10000)
-    .toString(16)
-    .substring(1)
-  ChunkActiveUploads[uuid] = {}
-
-  return res.json({
-    status: 'success',
-    data: {
-      session_id: uuid,
-      start_offset: 0,
-      end_offset: CHUNK_SIZE
-    }
-  })
-}
-
-const chunkUploadPart = (req, res) => {
-  setTimeout(() => {
-    const rand = Math.random()
-    if (rand <= 0.25) {
-      res.status(500)
-      res.json({ status: 'error', error: 'server' })
-    } else {
-      res.send({ status: 'success' })
-    }
-  }, 100 + parseInt(Math.random() * 2000, 10))
-}
-
-const chunkUploadFinish = (req, res) => {
-  setTimeout(() => {
-    const rand = Math.random()
-    if (rand <= 0.25) {
-      res.status(500)
-      res.json({ status: 'error', error: 'server' })
-    } else {
-      res.send({ status: 'success' })
-    }
-  }, 100 + parseInt(Math.random() * 2000, 10))
-}
-
-const chunkUpload = (req, res) => {
-  switch (req.body.phase) {
-    case 'start':
-      return chunkUploadStart(req, res)
-
-    case 'upload':
-      return chunkUploadPart(req, res)
-
-    case 'finish':
-      return chunkUploadFinish(req, res)
-  }
-}
 
 function baseConfig() {
   let config = {
@@ -242,6 +186,8 @@ module.exports = merge(baseConfig(), {
       let del = function (req, res) {
         res.json({ success: true })
       }
+
+      const bodyParser = require('webpack-body-parser')
 
       // Chunk upload
       app.post('/upload/chunk', bodyParser.json(), chunkUpload)
